+++ packages/mobile/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.3-alpha.0](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@2.0.2-alpha.0...@quiet/mobile@2.0.3-alpha.0) (2023-10-26)
+
+
+### Features
+
+* version code 366 ([cebf886](https://github.com/TryQuiet/quiet/commit/cebf886aee2c4ec6e0ef4e66219d9dc0afecb98c))
+
+
+
+
+
+## [2.0.2-alpha.0](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@2.0.1-alpha.7...@quiet/mobile@2.0.2-alpha.0) (2023-10-26)
+
+**Note:** Version bump only for package @quiet/mobile
+
+
+
+
+
 ## [2.0.1-alpha.7](https://github.com/TryQuiet/quiet/compare/@quiet/mobile@2.0.1-alpha.6...@quiet/mobile@2.0.1-alpha.7) (2023-10-16)
 
 **Note:** Version bump only for package @quiet/mobile